correções finais

+++ evasao_ifpb/lib/model/chart/pie_chart_model.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 class PieChartModel {
 
-  final int qtd;
+  final double qtd;
   final String nome;
   final Color? color;
